cli: Move profit-analysis commands to asset.rs and enhance detail output
- Move all profit-analysis handlers from trade.rs to asset.rs
- Enrich detail output: show holding value, buy/sell totals, fees,
  and itemized breakdowns for underlying and derivative
- Fix fund items missing symbol: fallback to isin field
